functional-tests: Fix configuration to work with Meson
Previously the functional-test suite would test the installed version of
Tracker, but would need to be run from a configured Tracker source+build tree.

I have changed how things are configured and now with Meson the functional tests
can be run completely from the build tree, without needing to install
anything. The Autotools `make functional-test` target works the same as
before.

Instead of generating a Python source file that contains the build-time
configuration, we write that to a .json file which is read when the
common.utils.configuration module is imported. This controls the
locations of the various things that the tests require, which allows us
to support any kind fo build tree layout.

With the new setup it should also be possible to add a target that
explicitly installs the functional tests, with a different
configuration.json file such that they test the installed version of
Tracker. That's not yet implemented.